requirements.txt
# Moving Accessibility Analyzer - Web Interface
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
altair>=5.0.0
Pillow>=10.0.0

# Chatbot dependencies
httpx>=0.25.0
python-dotenv>=1.0.0

# Core analyzer dependencies (ya las tienes)
# No necesitas reinstalar estas si ya funcionan:
# - No external dependencies needed for the core analyzer